Log and Dev Revamp
This commit changes developer module so that it is accessed like every other module as well as gaining a new developer function. The newest member of this family of admin system modules is the Log module, it is able to trace what file and line accesses each module as well as now supporting the refusal of certain file path locations accessing modules.  A file for global variables has also been made.