Command Module Complete
Finally completed the command module - this commit includes some serious bug fixes which includes but is not limited to: actually allowing more than one command to be registered at a time (quite an important one to fix), correcting the data type of the command codes (this is so that correct math functions can be executed with it), and allowing commands to be properly executed from the client and carried out serverside. This commit also comes with a small command testing file to show how it can be used.