Added The Permission Module
This module is done right out of the box as shown in the permission_module file in the JAAS folder. Permissions used to shared, now they are server side as they don't need to be on client for a majority of the time. As the user interface will be using permissions to show what people have access to, I might add some client side code to get if the client has access to something.