A Rank-tastic Commit
This commit finally finishes the rank removal functions. This has mean't all the modules that stores rank codes (Which is almost all the modules) now has to have a hook to update all rank codes when a rank code is removed. It supplies a function that executes that change in bits in complexity constant. The Command module also has a change, it now has four new data type it supports. All objects like a rank object now have metatables that show what type of object it is allowing less guess if the table is just a table or a JAAS object. New Commands have been added