Extended hook Interface
This just extends the hook interface a little bit to account for other uses. Although this in no way replaces the default hook system, it does a slightly different job. Whilst the default one allows for hook functions to return data so that it can affect the process, the JAAS hook system creates a new thread to run over the hook functions. This makes the two systems covers two different needs.