Module Upgrade
Added some new infrastructure and a new toy. Modules are a bit more automated with a post init handle for adding admin system features that can connect to module resources. SQL now has a new class that makes writing SQL queries a bit easier, not sure if they make it faster, this'll have to be tested.