Syntax Change and Addition
This commit shakes a few things up, some still need to be tested but most work. Changed the initialisation file - this includes altering the JAAS include function to make the code more pretty whilst also support table inputs and changing how its accessed, hook system now works by indexes so you can access current functions and set them using the Hook table, made it so the coroutines will actually resume the threads, changed the include loop to be more straight forward, and finishing it off with getting rid of wasteful code in the SharedSync function. Most of the changes to the other modules is to accomodate these changes so it doesn't get any syntax errors. The Player module now uses the Module handle so it doesn't require a rank library to run whilst the Rank Module now has a handle to prepare for a client sync of Rank data with a new permission to check if the player has access to seeing invisible ranks (Finally actually at a point where that single variable that has been there from the start actually does anything). The files that add the commands and permission are accommodating changes with permissions adding descriptions and removing the now gone Rank library that was needed. Rank related commands have been moved into the Rank module's Hande. New commands have been added as utility which now set the move types of players, not complicated.