SQL Change
This fixes the SQL module so it will now properly execute queries (forgot that you can't read the length of tables that aren't in order). Modified the player table so that it takes SteamID64 instead of text, should mean that it doesn't require an index. Also made it so SQL acts as a authentication for commands and permissions.