Rank, Command, and Permission Modification Channel
This commit adds modification channels to the main modules, instead of using commands to modify these important modules net messages will be used. This is so it can avoid using commands that may be confusing in context and use so using a user interface to be the primary form of modification is always preffered. Developer module also has a new class, the switch case which should make large if statements a bit faster as the sacrafice of memory. Tests will need to be done for speed, if if statements are found to be faster then it will replace all switch cases but for now they're quite useful (Assuming that they're faster). I assume they're faster because they're taking a O(1) approach to finding an answer instead of O(n).