Small Changes
Added a new command that kicks players. Added the argument builder as a typefunc so now functions can now be passed them. Modified the JAAS register slightly so it uses indexes instead of function parameters to add the the files. The JAAS register will also ErrorNoHalt instead of push an error if the file isn't found.