Finishing Log Module
The functions writing the logs are finished as well as the functions retrieving the logs for the client. Admin and Superadmin print functions have been removed because, well its mostly going to be used from the server. The file class has some new functions to support 64 bit unsigned and signed integers.