Log Expansion
Logs should now cover every basic action in the admin system. Going through the code has also allowed me to notice missing code as well as ways on improving former code. Chat logging now supports colours using a simple formating. Access Module now has a more indepth modification channel that allows not only the modification of Access Groups but the modification of Access Values of Permissions, Commands, and Ranks. I have replaced the default joining and disconnecting messages with JAAS messages which get automatically logged. I have moved the User commands to the Player Module so that it can use the same Log labels whilst also moving the test commands out and moving them with the rest of the commands. They all have log functions broadcasting to the admin system and the rest of the server what has been executed. Hidden logs have now become a thing where they should be recorded but would clob up the log viewer. Using the system it has been made clear that adding Logs is not the fastest thing or the most intuitive thing to do. I will look into a different way of going about it in the future to make it easier to use. Currently in this build not everything is working, writing in 64 bit is not perfect by the looks of the Hex reader but the reader I have built also isn't working so I have to fix that before I can fix 64 bit writing. There is also an error that occurs when levels are changed with the Command code syncing. The main bulk of it is working but a full and in-depth test will be performed before release to make sure there are no hidden errors or bugs.