Log Module Finish?
From what I can see at this point in time, the log module logic is done with the only thing I can see needing to be done in the future is playing around with the UI code. I have removed the 64 bit based code from the file class as whatever I did, it did not want to play ball so it now saves SteamIDs are strings which has been the only thing I've made for the file class that is constantly working. Code pulling log data now require a permission (I must of completely forgot to add that before), clients with the same permission are able to request all the dates that there are logs for which basically returns a list of file names. Logs can now be echoed to the client's console.  Handles now run as I have realised this whole time they haven't been executing and I have fixed the error that occured in the Command module about nil category names. The PlayerCanSeePlayersChat permission now uses the Log chat function. The commands in the Core Command module now use the proper class function for targetting as well as the function dropping the check if the Rank Library is the Rank Library as we know it is. The Rank module can now be requested for its list of ranks along with a permission for hiding the the invisible ranks. The command feedback hook now uses the correct table index. The biggest isse currently is the fact that when you join a game, it takes a few seconds for it to execute the lua. If this is a problem for staff members with more permissions or not, I would like to make executing the lua as quick as possible so a source will need to be tracked down and fixed.