New Module
As I'm working on the interface, I thought that I'd release the module that the interface uses. This is offically the smallest module with it only to allow a standard way of creating a interface for JAAS. Any interfaces must be added through this and can only support one interface at a time with the last to be registered being used.