Post JAAS Core Code
Renamed the Post JAAS register accordingly along with introducing the long awaited User Interface! I wanted to push this when it was actually finished as its been in development for months now but its time I finally pushed this. I've been having trouble with reading logs that have had Bots being saved as Players so an error has passed every time on reading, so code has been introduced to stop them being saved as a player and as an entity. For the interface I am using a library to make development of the interface easier and faster with the library handling Paint functions so I don't need to. The interface has a working command and log tabs with just Rank and Access to go (Hopefully by today).