Small Fixes
Serverside Command execution now recognises Player separately from String, even though it technically was doing that before. Fixed a bug with the log reader that stopped reading all log files once a hidden log was reached, will need to either remove hidden logs or fix it. Introduced a hook for the client that will be ran once the Rank Table is pulled so objects can be built from it, the Rank List is now working. Made the Add and Remove commands strictly server-side as Clients have the Rank Tab to properly add and remove players from Ranks.