Final Tab
The access tab is finally finished and has officially made the interface file the largest file in JAAS. The interface was a lot more complex that I initially thought with little planning and design going into actually supporting what I wanted out of the designs but it is finally finished. Not completely working as intended with commands being shown that you don't technically have access to and access groups having little affect over the other tabs. Whilst response can easily interact with the same its built to control, little connection with others which is obviously due to the order in which it was written with the first to last panels being written having the lowest to highest quality interface, this quality even seeps into the code with the Rank tab's code being all over the place and the Access tab's code being nicely organised due to it sharing similar code. I am very glad I built the interface as it taught me a lot about complex interface design, planning, and coding that sadly my course never actually taught me.  The future of this admin system will look different, this is not release ready at all with the serverside having sturdy interactions and clientside being built quickly with sticks and duct tape. The future would require a reworking of the interaction between client and server with data being automatically updated along with permissions being integrated into the updating process, the future of the interface will have an entire rework of the command tab, a literal delete and re-do. The rest of the tabs I am generally happy with, the only thing changing with the rank tab being permissions, commands and ranks being hidden automatically by their access groups. Whilst I am happy with most of the admin system, the shoddy nature of the client and the only way a player will ever interact with the admin system, makes the system feel cheap and fragile when it should feel anything but.