Noticing some small details
Imagine going through your code getting ready for its hand in and noticing some blatant code not working as it should without having to test it, well that is what the commit is. The Module file function should do an error as its not supposed to be executed, Rank iterator isn't flip flopping to returning nothing, and Access code check isn't accepting values below itself thus making Access Group not inheriting the values below it.