Rework Base Module
This creates a new focus for how modules can be built, this is mainly used as a concept introduction. One thing that hasn't been fully thought out is how code is going to be executed and the order in how its executed is going to be managed. This concept is shown fully in its final lines of code with Pre, Init, and Post functions separated in Client, Server, and Shared. One concept that hasn't been thought out is how data is going to be automatically updated, rather this file offering a method for minimising the code usage of automatic data updating between Client and Server. I've also introduced a new method for executing code after a certain module, rather than adding a function to a table, you declare that the module depends on another module and it creates a Post<Module_Name> function so that it can be executed at a later date after the module is executed.