Extended Base
Added quite a bit of code, mostly reworking the log module as well as how its accessed. Log files are now their own separate object that can be written and read on the net and file as well as placing the log record writing code within its NetWrite function rather than keep it locked behind a certain function. I felt like it would keep consistency between states, force developers to use the LogRecord object for pushing a new log. Also made the log registering function properly Shared so that index numbers can rise in tandem rather than just on the server side. Also introduced the first code to move data to the client on request. It is missing code to check if permissions exist which is a very important part of that code but I still need to fully decide on how that is going to happen.