Base Remix
Rewrote how the Object system works, I was wasn't sure how much of it really needs to exist specially when you cannot get the main benefit of stringing together inherited classes and when most of the classes need to be overwritten. So instead of an Object table, its an Object function that does the same thing as the New function. Most of the objects that inherited from it like the NetObject and FileObject are now gone with the only remnants being the names of the functions that will stick with current objects and any future ones. I've added some DO ... END blocks so that the editor will allow me to close sections of the code. I've decided that Module Access is going to work in a very similar way to the original way in that a function gets executed after everything is done executing, a slight change in that the function is apart of each Module. It does mean that there will only be one post function per module per state but if the module is organised nicely it shouldn't be a problem. There was some glaring obvious issues that I was able to fix on a second look through the code. Plus I have added the Permission Checks to two of the functions that needed it which means I only have one TODO but its questionable if it needs doing because the functions to pull the log file for the client has already been finished.