Adding Reworked Permission Module
Using the code from the Reworked Base, I have rewritten the Permission Module. Most of the module is the same, very similar logic with many of the functions having the exact same name. Only differences is that Permissions will always be synced between all players, be it if they have access to change the Permissions or not, all of the changes to the Permissions will be done with the Permission Object, and other module objects aren't automatically supported.  As you can see by the changes that have been made to the module and base, I feel it has streamlined the module hugely allowing a greater deal of Client access that gets automatically synced along with changes made to the server side. My only, lets say, concern is that every client has access to this data. I'm not sure if the client having access to data is in anyway a flaw but every change to permissions, every slight change and reverts backs, get sent to every single connected client.