Rewrote Initialisation File
This was kind of a waste of time as the starting code of the original Initialisation file was kind of perfect. The hooking system is complex and covers quite a lot, I'm generally quite happy with it. Only change has been that a ReadOnly function has been made for it that throws and error and the code to check the amount of hook functions present in the Remove section would not work due to the table using strings for keys, this would mean that removing a single hook would mean removing all of the hooks with how it is programmed. Obviously it is not finished, its missing the code to include modules and third party modules. Which will be decided upon at a later date. At this stage I just want to make sure that I liked how current hook system worked before installing it into the Reworked modules.