Organised and Trimmed Fat
Organised the code a bit along with trimming some unnecessary code that was created in the first phase of reworking the base code. Fixed some blaring mistakes that was missed the first time I went through it all. Obviously not testing the code makes it difficult to see blaring mistakes.