Access Group Rework
Introduces all the new techniques that have been established in the other reworked modules. I feel that the module is a lot cleaner than the original and easier to read. Only thing that is missing like the other modules post all the recent Commits is a serious lack of logs. The original ModifyChannel is far more complex than with far more support to the varying types of changes that could happen to an Access Group. Which is something I'm going to cover next before I move onto the next Module. A glaring change is the fact that this was the second module to be made, when originally this module was made last due to it being a last minute feature to JAAS. Due to it being integral to modifying channels I decided to do it after *the* most important module.