Modification Message and Protection
I've decided that the method for modifying module data is through Messages that tailor to setting and reading the messages. It bloats the modules hugely, line numbers have definitely increased from the original but I think its worth it for ease of use. A slight change is that read objects from the Net now do not automatically write into the database. I feel like I'm going to need to do some more playing around with the objects before the other modules are made.