Remove some Server code
This piece of code requires the Server side so it really needs to be removed for the client. I've decided that Access Groups require all to be pulled over before it can be viewed. Which is an obvious thought but really it can be programmed in a way to automatically pull on request.