Reworked Rank Module
This is probably the most finished module that has been reworked so far with it covering everything on the server and client with net code that automatically syncs data between server and client. I've copied the rank remove algorithm from the original module as its perfect but I've written a completely new and untested multi-remove algorithm. Its theoretically sound using the same techniques as the single remove algorithm with it viewing codes as sections to remove rather that single bits to which it will put together at the end. Whilst it is untested, I feel that it could work first time due to it using the exact same technique for capturing unchanged bits.