Reworked Player Module
This is probably the quickest module that I've had to rework which is not very surprising, really the original module was overdone. Biggest change from the original module is there is no longer a player module, mostly due to the fact that JAAS really did not need it with GMOD already providing a perfectly reasonable player object, so now functions are only available through that method.
There is a new column for the player, LastConnected. The idea behind this new data point is that it could serve as a fancy feature for the log reader that can show how long ago a player disconnected or a potential JAAS scoreboard could show recently disconnected players and how long ago they disconnected. Basically its a data point that does not already exist. Potentially another data point could exist for tracking the amount of time the player has spent on the server.
Other than the changes compared to the original, most of the techniques presented are pretty uniform compared with the reworked modules although this module could very well be awarded the smallest module even if these reworks have proven to be larger than the originals. Although its obvious the larger size will reduce the amount of code required to use this modules from a third party source or the interface which is the main purpose of the rework. One thing that was nice about this module was its size was easier to think through the Net Code. Which has proven to be the largest and most complex part of all of these modules and has frequently played with my train of thought. A method for organising Net Code really is needed so holes in my data and communication between states can become more obvious at first sight.