Net Code Checklist
I figured out a way to organise Net Code and that is to make a checklist and Do blocks with comments connecting back to the checklist. Just writing the checklist showed that I was missing a crucial part to the Net Code and I thought that the module was finished so already the checklist was doing what it was meant to. This will repeated for all existing and future modules (Really wish I had done this sooner)