Permission Net Code Checklist
Done the checklist for the Permission module, again I'm still so happy with this method. So simple but so effective to organising this and it makes me want to make one for the module itself but the other part of the code isn't as interweaving as the net code so I don't think its as necessary. One thing that I did change is the fact that the NetRead and NetWrite functions now return self, I've been coding it like it already supported it but now it actually does.