New Feature, Fixed Hooks, and Proper Maintenance
Biggest inclusion with this Commit and one that really needed to exist is a function that will check if a player's access to a Permission has changed. This definitely should not be used for all the permissions but a select few. A very useful function for the GUI that relies on dynamically changing based on what the player has access to. This adds so much functionality to the interface that it didn't have before.
Whilst creating the function, I double checked how to add hook functions, turns out its not an index but a function parameter that chooses the type of hook... soooo I had to change all the hook functions on all the modules.
The rule is if you store a code in a table or in anyway, it must be maintained through a Rank Remove Hook. Code maintenance is very important and because I did it once, I thought that its time for it to be included in all the modules so now all the modules have proper code maintenance.