Net Code Checklist
Last module to have its Net Code checked, I missed more here than I did on any of the modules so again, the checklist has been a life saver.
Due to the module's Net Code relying heavily on if the Client has access to viewing Access Groups, I thought it was time to introduce a function that checks a player's access to a permission dynamically, with it being created in the Player module I added its first use in this module where it checks if the LocalPlayer has access to viewing it. Something that would be completely impossible without the rework so its nice to see the rework already adding functionality beyond what the original offered.
Plus all the hook fixes and proper code maintenance.