Reworked Command Module
New module, this was made pretty damn quickly and finished when I should really be getting some sleep. With this Commit I only have one more module to complete and that module is THE most easiest module to complete and its stepping into the realm of the reason I did this in the first place; the interface. Biggest change from the original is the parameter builder has been fully fleshed out and comes with way more interface options than before like number types supporting min and max variables, removing the rank types, and supporting filters for the player types.
I'm going to say it again but that Net Code Checklist, greatest decision I've made, it makes coding Net Code sooo much easier. I'm going to use checklists for all of the complex coding situations this admin system puts me in.
One thing I do have to say though, programming this rework has been so much easier than programming the original due to the methods I have started in the other modules. It may be larger if counted by string but more of the lines are keywords that don't execute anything.