Finished GUI Module Rework
A simple module made a little more complex, using the new config system the choice of interface is chosen and can be overwritten in the Config.lua file.
I've added a few more features with interfaces being an object now and the OpenMenu and CloseMenu being functions on that object that need to be overwritten. Interface objects now come with the function OpenTab so you can select certain Tabs to open.
I've also brought over the ability to Register tabs, this can mean tab panels are going to be built in the modules but I feel like this is going to be modified the most whilst I create the interface.