Command Module Expansion
This module is now huge, I've changed my mind with how the keys for the command_table work so that they're two different tables and keys. Really I think its easier to think and work with when its two rather than a string concatenated together. Really wish I did this from the start and I'm hoping that I've caught all the times I use the old key.
I've added command flags, wish is just confusingly, only one flag so far which determines if the Command is a Server command only. I've added the JAAS objects as Parameter objects but they have less features than all of the other parameter objects as I'd like developers to avoid using them as use Tabs instead to work with JAAS objects, supporting them feels important so as to support commands that add players to Ranks from the dedicated server. Imagine if a staff member ruins the admin system permissions, throwing everyone off but themselves and leaving. The server owner would be basically screwed and be forced to reset the whole admin system to fix it.
I've also added support for executing commands from Console by creating a way to recognise a stream of characters as a certain object to convert to. I'm going to work to making it support an autocomplete function as well as support a chat command execute.