Added Logs
I've finished adding the logs to all the modules, decided to do it all in one Commit rather than one Commit for each module as I've continuously gone back and changed my mind.
One thing that I am realising that Logs have continued to be the least developer friendly part of JAAS. File logs are terribly slow to code and not obvious in how they're going to turn out. It doesn't support the various different objects of JAAS with ONLY Ranks being supported. I've obviously copied the old method for Logging to file and I didn't like it before, why did I keep it this time around when its the best time to make a huge change. I've obviously rushed the Log section and I've already written the code to write to file. It is a system that has proven to work but also something that I disliked the most to code. One thing that I can see as a huge problem is that in the module itself, the order in which Logs are registered cannot be altered in future updates due to any logs being written ahead of time losing complete connection to the order in which they should be built for the client viewer.
One thing that I have changed from the original is how coloured chat and console text is sent to the client, with their being three different permissions per module. Minor, Major, and console permission. Minor: Being small updates that superadmins would be interested in, Major: being the big updates that general admins would be interested in, and console updates that update the client with every single change that happens with the module. The console logs give more logs than anything other logging method. I've also introduced object and data type based colouring for the strings, a separate colour for each basic data type, and JAAS object.
I've also updated some of the objects whilst creating the logs have I have forgotten something quite important,  a default access check for all of the objects, now renamed Global Access for some clarity and the players have Default User.