Syntax Fix
Test finished and it wasn't that bad. First time I've done a syntax check post writing a huge amount of code without doing any the entire time. Arguably a bad idea but what has changed really was not much although it almost covering every module. There were very few syntax problems, with the biggest syntax error being how I use Hooks, the fact that I overwrite object code variable with the function creating the object for use, for some reason I never changed code to use the object function in Access Groups, the glaring fact that I've been using the wrong symbols to concatenate strings together which all needed to be changed, that I used the wrong sql function to execute a query in the CreateTable function in Base, I was using repeat wrong as it expects a true value to break whilst I thought it worked like a while loop and used a false value to break (This froze Gmod EVERY SINGLE TIME when joining the server). And really that is all of the syntax issues with everything that I wrote which comparatively with all of the code that I wrote, is really small.