GUI Extension
Added more code to the module and the objects to support more.
I've decided that a separate Panel file is not really needed, GUI is executed in the exact same place and the name gives the theme that covers everything GUI related for JAAS.
Work has started on a replacement for tdlib, I feel having this capability for JAAS is important for maintaining control over its own license plus increasing my own skill in this section of code that should allow me to have more control with how the JAAS interfaces look.
I've also created a tooltip function, if its required it another question entirely but I feel having great control on how it looks is something that would be useful.
As this module is also going to be acting as the Panel module from the old version, I've already created a Panel. Its untested but it should be an incredibly useful Panel even if the scope of its capability is much larger than it needs to be. My designs for the Panel is very complex and so far it would only support the main menu side bar. Supporting the dynamic visibility of objects, sorting of intended positions in the list, and animating the updated position movements.