Finished Interface Base
I've finished all the code to manage the tabs. I have to decide if I'm going to make the tabs in the file or in the module. I'm leaning towards making them here so the tabs can be per interface so other developers can experiment with their own versions of the same things rather than JAAS pushing its own default tab on all the interfaces with them using code to ignore certain tabs and recreating it themself.