Added Panel and expanded Parameters
I've added a Command ComboBox that will automatically fill with all current Commands. Added a function to iterate through commands. Added extra functions to the parameters objects and removed the ability to access the multi select parameters. Might think about adding them in the future but that is if people really wanted it.