Add Command and Rank Tab
Finally finished the Command and Rank tab. Three more tabs to go. Making these tabs does remind me why I don't like making UIs BUT it is nearly done. Although I feel, the worst is yet to come. The Access Group tab is a bit more complicated, a bit more of the same code as the Rank tab. I would like to create a custom control to control the order of inheritance of the Access Groups. I might change the design of the Log, I'm not completely sure, but for one thing, I'm definitely including the ability to filter logs plus the ability to hover over Players and bring up information.